+++ app/src/main/java/com/socialite/solite_pos/data/source/local/entity/room/master/Promo.kt
@@ -7,60 +7,42 @@ import androidx.room.PrimaryKey
 import androidx.sqlite.db.SimpleSQLiteQuery
 import com.socialite.solite_pos.data.source.local.room.AppDatabase
 import com.socialite.solite_pos.data.source.local.room.AppDatabase.Companion.UPLOAD
-import com.socialite.solite_pos.data.source.remote.response.entity.PromoResponse
 import com.socialite.solite_pos.view.ui.DropdownItem
 import java.io.Serializable
-import java.util.UUID
 
 @Entity(
-    tableName = Promo.DB_NAME,
-    indices = [
-        Index(value = [Promo.ID])
-    ]
+        tableName = Promo.DB_NAME,
+        indices = [
+            Index(value = [Promo.ID])
+        ]
 )
 data class Promo(
 
-    @PrimaryKey(autoGenerate = true)
-    @ColumnInfo(name = ID)
-    var id: Long,
+        @PrimaryKey(autoGenerate = true)
+        @ColumnInfo(name = ID)
+        var id: Long,
 
-    @ColumnInfo(name = AppDatabase.REPLACED_UUID, defaultValue = "")
-    val new_id: String,
+        @ColumnInfo(name = AppDatabase.REPLACED_UUID, defaultValue = "")
+        val new_id: String,
 
-    @ColumnInfo(name = NAME)
-    override var name: String,
+        @ColumnInfo(name = NAME)
+        override var name: String,
 
-    @ColumnInfo(name = DESC)
-    var desc: String,
+        @ColumnInfo(name = DESC)
+        var desc: String,
 
-    @ColumnInfo(name = CASH)
-    var isCash: Boolean,
+        @ColumnInfo(name = CASH)
+        var isCash: Boolean,
 
-    @ColumnInfo(name = VALUE)
-    var value: Int?,
+        @ColumnInfo(name = VALUE)
+        var value: Int?,
 
-    @ColumnInfo(name = STATUS)
-    var isActive: Boolean,
+        @ColumnInfo(name = STATUS)
+        var isActive: Boolean,
 
-    @ColumnInfo(name = UPLOAD)
-    var isUploaded: Boolean
+        @ColumnInfo(name = UPLOAD)
+        var isUploaded: Boolean
 ) : Serializable, DropdownItem {
-
-    fun isNewPromo() = id == 0L
-
-    fun isManualInput() = isCash && value == null
-
-    fun calculatePromo(total: Long, manualInput: Long?): Long {
-        return if (isCash) {
-            if (isManualInput()) {
-                manualInput ?: 0L
-            } else {
-                value?.toLong() ?: 0L
-            }
-        } else {
-            ((value!!.toFloat() / 100) * total).toLong()
-        }
-    }
 
     companion object {
         const val ID = "id_promo"
@@ -78,8 +60,8 @@ data class Promo(
             when (status) {
                 Status.ACTIVE -> {
                     query.append(" WHERE ")
-                        .append(STATUS)
-                        .append(" = ").append(status.code)
+                            .append(STATUS)
+                            .append(" = ").append(status.code)
                 }
 
                 else -> {
@@ -88,34 +70,6 @@ data class Promo(
             }
             return SimpleSQLiteQuery(query.toString())
         }
-
-        fun createNewPromo(
-            name: String,
-            desc: String,
-            isCash: Boolean,
-            value: Int?
-        ) = Promo(
-            id = 0L,
-            new_id = UUID.randomUUID().toString(),
-            name = name,
-            desc = desc,
-            isCash = isCash,
-            value = value,
-            isActive = true,
-            isUploaded = false
-        )
-    }
-
-    fun toResponse(): PromoResponse {
-        return PromoResponse(
-            id = id.toString(),
-            name = name,
-            desc = desc,
-            value = value,
-            isCash = isCash,
-            isActive = isActive,
-            isUploaded = true
-        )
     }
 
     enum class Status(val code: Int) {
