+++ app/src/main/java/com/socialite/solite_pos/data/source/local/room/PromosDao.kt
@@ -17,8 +17,8 @@ interface PromosDao {
     @RawQuery(observedEntities = [Promo::class])
     fun getPromos(query: SupportSQLiteQuery): Flow<List<Promo>>
 
-    @Query("SELECT * FROM ${Promo.DB_NAME} WHERE ${AppDatabase.UPLOAD} = 0")
-    suspend fun getNeedUploadPromos(): List<Promo>
+    @Query("SELECT * FROM ${NewPromo.DB_NAME} WHERE ${AppDatabase.UPLOAD} = 0")
+    suspend fun getNeedUploadPromos(): List<NewPromo>
 
     @RawQuery(observedEntities = [NewPromo::class])
     fun getNewPromos(query: SupportSQLiteQuery): Flow<List<NewPromo>>
@@ -33,7 +33,7 @@ interface PromosDao {
     suspend fun insertNewPromo(data: NewPromo)
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertPromos(data: List<Promo>)
+    suspend fun insertPromos(data: List<NewPromo>)
 
     @Update
     suspend fun updatePromo(data: Promo)
