+++ app/src/main/java/com/socialite/solite_pos/data/source/repository/impl/PromosRepositoryImpl.kt
@@ -35,7 +35,7 @@ class PromosRepositoryImpl(
     }
 
     override suspend fun insertPromo(data: NewPromo) = dao.insertNewPromo(data)
-    override suspend fun insertPromos(list: List<Promo>)  = dao.insertPromos(list)
+    override suspend fun insertPromos(list: List<NewPromo>)  = dao.insertPromos(list)
     override suspend fun updatePromo(data: NewPromo) = dao.updateNewPromo(data)
     override suspend fun getNeedUploadPromos() = dao.getNeedUploadPromos()
     override fun getPromos(query: SupportSQLiteQuery) = dao.getNewPromos(query)
